Improved showcase
- Added section concerning open source applications to showcase
- moved lists of apps into own files
- automatically sort the apps by name
