+++ website/src/react-native/showcase/openSource.js
@@ -1,0 +1,41 @@
+var openSource = [
+  {
+    name: 'HackerNews',
+    icon: 'http://a4.mzstatic.com/us/r30/Purple49/v4/86/90/b3/8690b33f-f8a4-b13d-e2cb-d6522653293e/icon175x175.jpeg',
+    link: 'https://github.com/iSimar/HackerNews-React-Native',
+    author: 'Simar',
+  }, {
+    name: 'ZhiHuDaily',
+    icon: 'https://raw.githubusercontent.com/race604/ZhiHuDaily-React-Native/master/ios/RCTZhiHuDaily/Images.xcassets/AppIcon.appiconset/ic_launcher.png',
+    link: 'https://github.com/race604/ZhiHuDaily-React-Native',
+    author: 'race604',
+  }, {
+    name: 'Finance',
+    icon: 'https://lh3.googleusercontent.com/3VIk0kX6WyaHdQkNJVdFqW6Kn-O-rZ1RUwvu1tjQamRC33ThMIGh3dkRALpSRPjUIw=w300-rw',
+    link: 'https://github.com/7kfpun/FinanceReactNative',
+    author: '7kfpun',
+  }, {
+    name: 'Ziliun',
+    icon: 'https://lh3.googleusercontent.com/c6ot13BVlU-xONcQi-llFmKXZCLRGbfrCv1RnctWtOELtPYMc0A52srXAfkU897QIg=w300-rw',
+    link: 'https://github.com/sonnylazuardi/ziliun-react-native',
+    author: 'Sonny Lazuardi',
+  }, {
+    name: 'GitFeed',
+    icon: 'http://a3.mzstatic.com/us/r30/Purple69/v4/00/28/b8/0028b895-dda8-1def-c879-4da2ceeef2dc/icon175x175.png',
+    link: 'https://github.com/xiekw2010/react-native-gitfeed',
+    author: 'David Tse',
+  }, {
+    name: 'No Fluff: Hiragana',
+    icon: 'https://lh3.googleusercontent.com/kStXwjpbPsu27E1nIEU1gfG0I8j9t5bAR_20OMhGZvu0j2vab3EbBV7O_KNZChjflZ_O',
+    link: 'https://play.google.com/store/apps/details?id=com.hiragana',
+    author: 'Matthias Sieber',
+    source: 'https://github.com/manonthemat/no-fluff-hiragana'
+  }, {
+    name: 'F8',
+    icon: 'http://a4.mzstatic.com/eu/r30/Purple69/v4/14/aa/49/14aa49ba-0c42-c7f4-ce2a-3945c4bfe5eb/icon175x175.jpeg',
+    link: 'https://github.com/fbsamples/f8app',
+    author: 'Facebook'
+  }
+];
+
+module.exports = openSource;
